remove cop code for now
