+++ tests/scai_examples.py
@@ -1,0 +1,92 @@
+import numpy as np
+
+
+def _parse_example_mask(s):
+    return np.array([[float(c) for c in line] for line in s.strip().splitlines()])
+
+
+EXAMPLES = {}
+
+EXAMPLES["empty"] = [
+    _parse_example_mask(
+        """
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+"""
+    ),
+    float("nan"),
+]
+
+EXAMPLES["2a"] = [
+    _parse_example_mask(
+        """
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000001110000
+00000000000100110000
+00000000000000000000
+00000001111110000000
+00000001111110000000
+00000000000001000000
+00000000000100000000
+00000000000000000000
+00000000000000000000
+"""
+    ),
+    5.42236,
+]
+
+EXAMPLES["2b"] = [
+    _parse_example_mask(
+        """
+00000011100000000000
+00000011111000000000
+00000000110000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000111110000
+00000000000110000000
+00000000000000000001
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00000000000000000000
+00100000000000000000
+00000000000000000000
+00000000000000000000
+00000000010000000000
+00000000010000000000
+"""
+    ),
+    16.7,
+]
